+++ frontend/dashboard/public/manifest.json
