+++ frontend/dashboard/src/DashboardComponent.js
