+++ frontend/ingestion/src/thp_logo.png
