+++ src/dashboard/generate_snapshots.py
@@ -5,11 +5,11 @@ File to generate stratified reports and tests.
 import logging
 import pandas as pd
 import warnings
-from src.config_manager import load_config
-from src.metrics import generate_report
-from src.tests import generate_tests
-from src.etl import etl_pipeline
-from src.stratify import DataSplitter
+from src.utils.config_manager import load_config
+from src.monitoring.metrics import generate_report
+from src.monitoring.tests import generate_tests
+from src.data_preprocessing.etl import etl_pipeline
+from src.monitoring.stratify import DataSplitter
 from sklearn.exceptions import UndefinedMetricWarning
 from datetime import datetime
 from scripts.data_details import load_details
