+++ src/dashboard/workspace_manager.py
