+++ src/data_preprocessing/etl.py
@@ -3,8 +3,8 @@ ETL pipeline script. This script is responsible for loading, validating, and spl
 """
 
 import os
-from ingestion.scripts.fetch_data import fetch_and_merge
-from src.validate import validate_data
+from src.data_preprocessing.fetch_data import fetch_and_merge
+from src.data_preprocessing.validate import validate_data
 from scripts.data_details import data_details
 import pandas as pd
 import logging
